Initial upload
So... I just messed up. I don't know how I caused my code not to work, but I did. Now, I have to dig through my code and find out where I messed up instead of reverting to an older commit. Serves me right for not using github from the start.
The problem isn't that bad though. Curses opens the wrapper function, and the border box is drawn. However, the icons for files, folders, etc. are not drawn on the screen. As far as I can remember, I wrapped them in a loop, and after that, the icons no longer displayed. I'll see if I can get it back to how it was working before for my next commit.