Read documentation
So it turns out that I just needed to finish reading the documentation on the getkey() function. It waits for a keypress to allow the code to continue, and so was stopping the rest of the code in my while loop (that renders the icons) from running. The constant refreshes had nothing to do with it!
I'm going to tweak the code a little more so that it runs getkey() at the tail end of the loop, and passes its result on to the next cycle of the loop.