Added more logging statements
Added more log statements to the while loop on line 191 to make it more clear what is happening in the program.
This is the last commit before I rip out and re-do the entire loop. My goal is to use an array of objects instead of piecing together the text and icon haphazardly.